split up the ConfigureMaster phase
