+++ lib/pharos/cluster_manager.rb
@@ -74,6 +74,9 @@ module Pharos
       end
 
       apply_phase(Phases::ConfigureSecretsEncryption, master_hosts, ssh: true, parallel: false)
+      apply_phase(Phases::SetupMaster, master_hosts, ssh: true, parallel: true)
+      apply_phase(Phases::UpgradeMaster, master_hosts, ssh: true, parallel: false)
+      apply_phase(Phases::ConfigureKubelet, master_hosts, ssh: true, parallel: true)
       apply_phase(Phases::ConfigureMaster, master_hosts, ssh: true, parallel: false)
       apply_phase(Phases::MigrateWorker, config.worker_hosts, ssh: true, parallel: true, master: master_hosts.first)
       apply_phase(Phases::ConfigureKubelet, config.worker_hosts, ssh: true, parallel: true)
